# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NadlanConnect France API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

/**
 * Type de location (nu = Jeanbrun, meuble = LMNP)
 */
export type SimulationInputType = typeof SimulationInputType[keyof typeof SimulationInputType];


export const SimulationInputType = {
  nu: 'nu',
  meuble: 'meuble',
} as const;

export interface SimulationInput {
  /** Prix du bien en euros */
  prix: number;
  /** Loyer mensuel estimé en euros */
  loyer: number;
  /** Tranche marginale d'imposition (11, 30, 41, 45) */
  tmi: number;
  /** Type de location (nu = Jeanbrun, meuble = LMNP) */
  type: SimulationInputType;
  /** Charges annuelles en euros */
  charges: number;
  /** Durée de détention souhaitée en années */
  duree: number;
}

export interface SimulationResult {
  rendement_brut: string;
  rendement_net: string;
  gain_fiscal_annuel: number;
  economies_totales: number;
  resultat_fiscal: number;
  recommandation: string;
  score_investissement: number;
  points_vigilance: string[];
}

export interface PdfAnalyzeInput {
  /** Nom du fichier PDF uploadé */
  filename: string;
}

export interface PdfAnalysisResult {
  /** @nullable */
  prix: string | null;
  /** @nullable */
  surface: string | null;
  clauses: string[];
  risques: string[];
  resume: string;
}

export interface ApiError {
  error: string;
}

export interface AnthropicConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface AnthropicMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface AnthropicConversationInput {
  title: string;
}

export interface AnthropicMessageInput {
  content: string;
}

export interface AnthropicConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: AnthropicMessage[];
}

export interface AnthropicError {
  error: string;
}