# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: NadlanConnect France API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /simulate:
    post:
      operationId: simulate
      tags: [simulate]
      summary: Simulation IA d'investissement immobilier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SimulationInput"
      responses:
        "200":
          description: Résultat de la simulation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimulationResult"
        "400":
          description: Données invalides
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "500":
          description: Erreur serveur
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /pdf/analyze:
    post:
      operationId: analyzePdf
      tags: [pdf]
      summary: Analyse d'un document PDF par IA
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/PdfAnalyzeInput"
      responses:
        "200":
          description: Résultat de l'analyse PDF
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PdfAnalysisResult"
        "400":
          description: Fichier invalide
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "500":
          description: Erreur serveur
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
components:
  schemas:
...
    SimulationInput:
      type: object
      required:
        - prix
        - loyer
        - tmi
        - type
        - charges
        - duree
      properties:
        prix:
          type: number
          description: Prix du bien en euros
        loyer:
          type: number
          description: Loyer mensuel estimé en euros
        tmi:
          type: number
          description: Tranche marginale d'imposition (11, 30, 41, 45)
        type:
          type: string
          enum: [nu, meuble]
          description: Type de location (nu = Jeanbrun, meuble = LMNP)
        charges:
          type: number
          description: Charges annuelles en euros
        duree:
          type: number
          description: Durée de détention souhaitée en années
    SimulationResult:
      type: object
      required:
        - rendement_brut
        - rendement_net
        - gain_fiscal_annuel
        - economies_totales
        - resultat_fiscal
        - recommandation
        - score_investissement
        - points_vigilance
      properties:
        rendement_brut:
          type: string
        rendement_net:
          type: string
        gain_fiscal_annuel:
          type: number
        economies_totales:
          type: number
        resultat_fiscal:
          type: number
        recommandation:
          type: string
        score_investissement:
          type: number
        points_vigilance:
          type: array
          items:
            type: string
    PdfAnalyzeInput:
      type: object
      required:
        - filename
      properties:
        filename:
          type: string
          description: Nom du fichier PDF uploadé
    PdfAnalysisResult:
      type: object
      required:
        - prix
        - surface
        - clauses
        - risques
        - resume
      properties:
        prix:
          type: ["string", "null"]
        surface:
          type: ["string", "null"]
        clauses:
          type: array
          items:
            type: string
        risques:
          type: array
          items:
            type: string
        resume:
          type: string
    ApiError:
      type: object
      required:
        - error
      properties:
        error:
          type: string
...