# lib/api-zod/src/generated/types/pdfAnalysisResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NadlanConnect France API
 * OpenAPI spec version: 0.1.0
 */

export interface PdfAnalysisResult {
  /** @nullable */
  prix: string | null;
  /** @nullable */
  surface: string | null;
  clauses: string[];
  risques: string[];
  resume: string;
}