# lib/api-zod/src/generated/types/pdfAnalyzeInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NadlanConnect France API
 * OpenAPI spec version: 0.1.0
 */

export interface PdfAnalyzeInput {
  /** Nom du fichier PDF uploadé */
  filename: string;
}