# lib/api-zod/src/generated/types/simulationInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NadlanConnect France API
 * OpenAPI spec version: 0.1.0
 */
import type { SimulationInputType } from './simulationInputType';

export interface SimulationInput {
  /** Prix du bien en euros */
  prix: number;
  /** Loyer mensuel estimé en euros */
  loyer: number;
  /** Tranche marginale d'imposition (11, 30, 41, 45) */
  tmi: number;
  /** Type de location (nu = Jeanbrun, meuble = LMNP) */
  type: SimulationInputType;
  /** Charges annuelles en euros */
  charges: number;
  /** Durée de détention souhaitée en années */
  duree: number;
}