# lib/api-zod/src/generated/types/simulationResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NadlanConnect France API
 * OpenAPI spec version: 0.1.0
 */

export interface SimulationResult {
  rendement_brut: string;
  rendement_net: string;
  gain_fiscal_annuel: number;
  economies_totales: number;
  resultat_fiscal: number;
  recommandation: string;
  score_investissement: number;
  points_vigilance: string[];
}